20231114_fantastic_tool_and_how_to_use_them/ShadowPad Light/requirements.txt
capstone==5.0.1
cffi==1.16.0
cryptography==41.0.5
hexdump==3.3
pycparser==2.21
python-snappy==0.6.1
unicorn==2.0.1.post1
